cart context processors

+++ cart/cart.py
@@ -55,7 +55,7 @@ class Cart(object):
         return sum(item['quantity'] for item in self.cart.values())
 
     def get_total_price(self):
-        return sum(item['price'] * item['quantity'] for item in self.cart.values())
+        return sum(Decimal(item['price'] * item['quantity']) for item in self.cart.values())
 
     def clear(self):
         del self.session[settings.CART_SESSION_ID]
